feat: code deduplication
