+++ crates/block/Cargo.toml
@@ -11,6 +11,7 @@ path = "src/lib.rs"
 [features]
 default = ["net"]
 prover = []
+test-utils = ["dep:reth-storage-api"]
 net = [
     "dep:alethia-reth-primitives",
     "alethia-reth-primitives?/net",
@@ -43,6 +44,7 @@ reth-primitives = { workspace = true }
 reth-primitives-traits = { workspace = true }
 reth-revm = { workspace = true }
 reth-rpc-eth-api = { workspace = true, optional = true }
+reth-storage-api = { workspace = true, optional = true }
 reth-storage-errors = { workspace = true }
 reth-transaction-pool = { workspace = true, features = ["test-utils"] }
 revm-database-interface = { workspace = true }
