+++ crates/block/src/tx_selection/mod.rs
@@ -276,30 +276,27 @@ where
 mod tests {
     use std::sync::Arc;
 
-    use alloy_consensus::{Signed, TxLegacy};
-    use alloy_evm::{EvmEnv, EvmFactory};
-    use alloy_hardforks::ForkCondition;
-    use alloy_primitives::{B256, Bytes, ChainId, Signature, TxKind, U256};
-    use reth_ethereum::EthPrimitives;
-    use reth_evm::{
-        block::{BlockExecutionError, BlockExecutor},
-        execute::{BlockBuilder, BlockBuilderOutcome, ExecutorTx},
-    };
+    use alloy_evm::EvmFactory;
+    use alloy_primitives::Address;
+    use reth_evm::block::BlockExecutor;
     use reth_evm_ethereum::RethReceiptBuilder;
-    use reth_primitives::Recovered;
-    use reth_revm::{
-        State,
-        context::result::ExecutionResult,
-        db::InMemoryDB,
-        state::{AccountInfo, Bytecode, bytecode::opcode},
-    };
-    use reth_storage_api::StateProvider;
+    use reth_revm::State;
     use reth_transaction_pool::{TransactionOrigin, TransactionPool, test_utils::testing_pool};
 
     use super::{SelectionOutcome, TxSelectionConfig, select_and_execute_pool_transactions};
-    use crate::{executor::TaikoBlockExecutor, factory::TaikoBlockExecutionCtx};
-    use alethia_reth_chainspec::{TAIKO_DEVNET, hardfork::TaikoHardfork, spec::TaikoChainSpec};
-    use alethia_reth_evm::{factory::TaikoEvmFactory, spec::TaikoSpecId};
+    use crate::{
+        executor::TaikoBlockExecutor,
+        testutil::{
+            BENCH_LIMIT_TARGET, BENCH_SUCCESS_TARGET, ExecutorBackedBuilder, db_with_contracts,
+            recovered_tx, uzen_chain_spec, uzen_evm_env, uzen_execution_ctx,
+        },
+    };
+    use alethia_reth_evm::factory::TaikoEvmFactory;
+
+    const BENCH_INVALID_CALLER: Address = Address::with_last_byte(0x30);
+    const BENCH_INCLUDED_CALLER: Address = Address::with_last_byte(0x31);
+    const BENCH_LIMIT_CALLER: Address = Address::with_last_byte(0x32);
+    const BENCH_LATE_CALLER: Address = Address::with_last_byte(0x33);
 
     #[test]
     fn tx_selection_breaks_on_uzen_zk_gas_error_but_keeps_skipping_invalid_txs() {
@@ -371,182 +368,6 @@ mod tests {
             *lists[0].transactions[0].tx.tx_hash(),
             *recovered_tx(BENCH_INCLUDED_CALLER, BENCH_SUCCESS_TARGET, 0, 30).tx_hash()
         );
-        assert_eq!(builder.executor().receipts().len(), 1);
+        assert_eq!(builder.executor.receipts().len(), 1);
     }
-
-    struct ExecutorBackedBuilder<E> {
-        executor: E,
-    }
-
-    impl<E> BlockBuilder for ExecutorBackedBuilder<E>
-    where
-        E: BlockExecutor<
-                Transaction = reth_ethereum::TransactionSigned,
-                Receipt = reth_ethereum::Receipt,
-            >,
-    {
-        type Primitives = EthPrimitives;
-        type Executor = E;
-
-        fn apply_pre_execution_changes(&mut self) -> Result<(), BlockExecutionError> {
-            self.executor.apply_pre_execution_changes()
-        }
-
-        fn execute_transaction_with_commit_condition(
-            &mut self,
-            tx: impl ExecutorTx<Self::Executor>,
-            f: impl FnOnce(
-                &ExecutionResult<
-                    <<Self::Executor as BlockExecutor>::Evm as reth_evm::Evm>::HaltReason,
-                >,
-            ) -> reth_evm::block::CommitChanges,
-        ) -> Result<Option<u64>, BlockExecutionError> {
-            let tx = tx.into_parts();
-            self.executor.execute_transaction_with_commit_condition(tx, f)
-        }
-
-        fn finish(
-            self,
-            _state_provider: impl StateProvider,
-        ) -> Result<BlockBuilderOutcome<Self::Primitives>, BlockExecutionError> {
-            unreachable!("finish is not used in these unit tests")
-        }
-
-        fn executor_mut(&mut self) -> &mut Self::Executor {
-            &mut self.executor
-        }
-
-        fn executor(&self) -> &Self::Executor {
-            &self.executor
-        }
-
-        fn into_executor(self) -> Self::Executor {
-            self.executor
-        }
-    }
-
-    fn uzen_chain_spec() -> TaikoChainSpec {
-        let mut chain_spec = (*TAIKO_DEVNET).as_ref().clone();
-        chain_spec.inner.hardforks.insert(TaikoHardfork::Uzen, ForkCondition::Timestamp(0));
-        chain_spec
-    }
-
-    fn uzen_evm_env() -> EvmEnv<TaikoSpecId> {
-        let mut env: EvmEnv<TaikoSpecId> = EvmEnv::default();
-        env.cfg_env.spec = TaikoSpecId::UZEN;
-        env.cfg_env.chain_id = 167;
-        env.block_env.number = U256::from(1_u64);
-        env.block_env.timestamp = U256::from(1_u64);
-        env.block_env.gas_limit = 30_000_000;
-        env
-    }
-
-    fn uzen_execution_ctx<'a>() -> TaikoBlockExecutionCtx<'a> {
-        TaikoBlockExecutionCtx {
-            parent_hash: B256::ZERO,
-            parent_beacon_block_root: None,
-            ommers: &[],
-            withdrawals: None,
-            basefee_per_gas: 0,
-            extra_data: Bytes::default(),
-            is_uzen: true,
-            expected_difficulty: None,
-            finalized_block_zk_gas: Default::default(),
-        }
-    }
-
-    fn recovered_tx(
-        caller: alloy_primitives::Address,
-        to: alloy_primitives::Address,
-        nonce: u64,
-        gas_price: u64,
-    ) -> Recovered<reth_ethereum::TransactionSigned> {
-        let tx = TxLegacy {
-            chain_id: Some(ChainId::from(167_u64)),
-            nonce,
-            gas_price: gas_price.into(),
-            gas_limit: 5_000_000,
-            to: TxKind::Call(to),
-            value: U256::ZERO,
-            input: Bytes::default(),
-        };
-        let signature = Signature::new(U256::from(1_u64), U256::from(2_u64), false);
-        Recovered::new_unchecked(
-            Signed::new_unchecked(tx, signature, B256::with_last_byte(caller.as_slice()[19]))
-                .into(),
-            caller,
-        )
-    }
-
-    fn db_with_contracts(accounts: &[(alloy_primitives::Address, u64)]) -> InMemoryDB {
-        let mut db = InMemoryDB::default();
-        insert_contract(&mut db, BENCH_SUCCESS_TARGET, arithmetic_bytecode());
-        insert_contract(&mut db, BENCH_LIMIT_TARGET, limit_exceeding_keccak_bytecode());
-        for &(address, nonce) in accounts {
-            db.insert_account_info(
-                address,
-                AccountInfo {
-                    nonce,
-                    balance: U256::from(10_000_000_000_u64),
-                    ..Default::default()
-                },
-            );
-        }
-        db
-    }
-
-    fn insert_contract(
-        db: &mut InMemoryDB,
-        address: alloy_primitives::Address,
-        bytecode: Bytecode,
-    ) {
-        let code_hash = bytecode.hash_slow();
-        db.insert_account_info(
-            address,
-            AccountInfo {
-                nonce: 1,
-                balance: U256::ZERO,
-                code_hash,
-                code: Some(bytecode),
-                ..Default::default()
-            },
-        );
-    }
-
-    fn arithmetic_bytecode() -> Bytecode {
-        Bytecode::new_raw(Bytes::from(vec![
-            opcode::PUSH1,
-            0x01,
-            opcode::PUSH1,
-            0x02,
-            opcode::ADD,
-            opcode::STOP,
-        ]))
-    }
-
-    fn limit_exceeding_keccak_bytecode() -> Bytecode {
-        Bytecode::new_raw(Bytes::from(vec![
-            opcode::PUSH1,
-            0x20,
-            opcode::PUSH3,
-            0x10,
-            0x00,
-            0x00,
-            opcode::KECCAK256,
-            opcode::STOP,
-        ]))
-    }
-
-    const BENCH_INVALID_CALLER: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x30);
-    const BENCH_INCLUDED_CALLER: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x31);
-    const BENCH_LIMIT_CALLER: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x32);
-    const BENCH_LATE_CALLER: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x33);
-    const BENCH_SUCCESS_TARGET: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x21);
-    const BENCH_LIMIT_TARGET: alloy_primitives::Address =
-        alloy_primitives::Address::with_last_byte(0x22);
 }
